Modified   Config/DefaultEngine.ini Modified   Source/ProjectX/Core/AC_Inventory.cpp Modified   Source/ProjectX/Core/AC_Inventory.h Modified   Source/ProjectX/Core/BulletBase.cpp Modified   Source/ProjectX/Core/CharacterBase.cpp Modified   Source/ProjectX/Core/CharacterBase.h Modified   Source/ProjectX/Core/EnemyBase.cpp Modified   Source/ProjectX/Core/Interfaces/IWeapon.h Modified   Source/ProjectX/Core/WeaponBase.cpp Modified   Source/ProjectX/Core/WeaponBase.h Modified   Source/ProjectX/Datas/Weapons/WeaponsEnum.h Modified   Source/ProjectX/Notifies/AnimNotify_Holster.cpp Modified   Source/ProjectX/Notifies/AnimNotify_Holster.h Modified   Source/ProjectX/Weapons/Assault_Rifle_01.cpp Modified   Source/ProjectX/Weapons/Assault_Rifle_01.h Modified   Source/ProjectX/Weapons/Assault_Rifle_02.cpp Modified   Source/ProjectX/Weapons/Assault_Rifle_02.h Modified   Source/ProjectX/Weapons/Handgun_01.cpp Modified   Source/ProjectX/Weapons/Handgun_01.h Modified   Source/ProjectX/Weapons/Sniper_02.cpp Modified   Source/ProjectX/Weapons/Sniper_02.h Modified   Source/ProjectX/Widget/MainGameInterface.cpp Modified   Source/ProjectX/Widget/MainGameInterface.h Modified   Source/ProjectX/Widget/WeaponWidget.cpp Modified   Source/ProjectX/Widget/WeaponWidget.h Added      Source/ProjectX/Bullet/HandgunBullet.cpp Added      Source/ProjectX/Bullet/HandgunBullet.h Added      Source/ProjectX/Bullet/SniperBullet.cpp Added      Source/ProjectX/Bullet/SniperBullet.h Added      Source/ProjectX/Casing/SniperCasing.cpp Added      Source/ProjectX/Casing/SniperCasing.h Added      Source/ProjectX/Core/AmmoBoxCore.cpp Added      Source/ProjectX/Core/AmmoBoxCore.h Added      Source/ProjectX/Core/Interfaces/IInteractable.cpp Added      Source/ProjectX/Core/Interfaces/IInteractable.h Added      Source/ProjectX/Objects/AmmoBox/BigBulletAmmoBox.cpp Added      Source/ProjectX/Objects/AmmoBox/BigBulletAmmoBox.h Added      Source/ProjectX/Objects/AmmoBox/SmallBulletAmmoBox.cpp Added      Source/ProjectX/Objects/AmmoBox/SmallBulletAmmoBox.h Added      Source/ProjectX/Objects/Switch.cpp Added      Source/ProjectX/Objects/Switch.h Added      Source/ProjectX/Objects/Target_Hanged.cpp Added      Source/ProjectX/Objects/Target_Hanged.h Added      Source/ProjectX/Widget/InteractTextWidget.cpp Added      Source/ProjectX/Widget/InteractTextWidget.h

+++ Source/ProjectX/Widget/WeaponWidget.cpp
@@ -6,7 +6,7 @@
 #include "Core/Interfaces/IWeapon.h"
 #include "Utilities/Global.h"
 
-void UWeaponWidget::SetMaxAmmo(uint8 RemainAmmo)
+void UWeaponWidget::SetMaxAmmo(uint32 RemainAmmo)
 {
 //     int8 maxAmmo = Weapon->GetMaxAmmo();
     MaxAmmo->SetText(FText::AsNumber(RemainAmmo));
